refactor(tests): Restructure test organisation slightly
